+++ app/dev/dashboard/app-forms/AddAppForm.jsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import './forms.css';
+import '../forms.css';
 import { addApp, getBrandsForUser } from '@/app/db/database';
 import FileInput from './FileInput';
 import FormSubmitFeedback from './FormSubmitFeedback';
